requirements.txt
gpxpy
shapely
pyproj
requests
tqdm
folium
pyyaml
pandas
